WIP iOS contact picker implementation.
