+++ src/MobileKidsIdApp/MobileKidsIdApp/MobileKidsIdApp.iOS/Services/ContactPicker.cs
@@ -9,6 +9,7 @@ using ContactsUI;
 using Contacts;
 using Foundation;
 using UIKit;
+using System.Threading;
 
 [assembly: Xamarin.Forms.Dependency(typeof(MobileKidsIdApp.iOS.Services.ContactPicker))]
 namespace MobileKidsIdApp.iOS.Services
@@ -21,48 +22,85 @@ namespace MobileKidsIdApp.iOS.Services
     /// </remarks>
     public class ContactPicker :  IContactPicker
     {
-
-        protected ContactInfo SelectedContactInfo { get; set; }
-
-        public async Task<ContactInfo> GetSelectedContactInfo()
+        public Task<ContactInfo> GetSelectedContactInfo()
         {
-           
+            var picker = new CNContactPickerViewController()
+            {
+                DisplayedPropertyKeys = new NSString[] { CNContactKey.GivenName, CNContactKey.FamilyName }
+            };
+            
+            // Set up Contact Picker Delegate, TaskCompletionSource wrapper.
             var tcs = new TaskCompletionSource<ContactInfo>();
-
-            var picker = new CNContactPickerViewController();
-
-            // Select property to pick
-            picker.DisplayedPropertyKeys = new NSString[] { CNContactKey.GivenName, CNContactKey.FamilyName };
-
-            // Respond to selection
             var pickerDelegate = new ContactPickerDelegate();
             picker.Delegate = pickerDelegate;
 
-            pickerDelegate.ContactSelected += (contact) => {
-                SelectedContactInfo = new ContactInfo { Id = contact.Identifier, DisplayName = string.Format("{0} {1}", contact.GivenName, contact.FamilyName) };
+            pickerDelegate.ContactSelected += (contact) =>
+            {
+                tcs.SetResult(new ContactInfo { Id = contact.Identifier, FamilyName = contact.FamilyName, AdditionalName = contact.MiddleName, GivenName = contact.GivenName });
             };
 
+            // Display as modal dialog on main view.
             UIWindow window = UIApplication.SharedApplication.KeyWindow;
             UIViewController viewController = window.RootViewController;
             if (viewController == null)
             {
                 while (viewController.PresentedViewController != null)
                     viewController = viewController.PresentedViewController;
-
             }
 
             viewController.PresentViewController(picker, true, null);
 
-
-            return SelectedContactInfo;
-            // Display picker
-
-   
+            return tcs.Task;
             
+        }
+    }
+
+    internal class ContactPickerDelegate : CNContactPickerDelegate
+    {
+        #region Constructors
+        public ContactPickerDelegate() { }
+
+        public ContactPickerDelegate(IntPtr handle) : base(handle)
+        {
+        }
+
+        protected ContactPickerDelegate(NSObjectFlag t) : base(t)
+        {
+        }
+        #endregion
+
+        public override void DidSelectContact(CNContactPickerViewController picker, CNContact contact)
+        {
+            // Raise the contact selected event
+            RaiseContactSelected(contact);
         }
 
 
-    }
+        #region Events
+        public delegate void SelectionCanceledDelegate();
+        public event SelectionCanceledDelegate SelectionCanceled;
 
+        internal void RaiseSelectionCanceled()
+        {
+            if (this.SelectionCanceled != null) this.SelectionCanceled();
+        }
+
+        public delegate void ContactSelectedDelegate(CNContact contact);
+        public event ContactSelectedDelegate ContactSelected;
+
+        internal void RaiseContactSelected(CNContact contact)
+        {
+            if (this.ContactSelected != null) this.ContactSelected(contact);
+        }
+
+        public delegate void ContactPropertySelectedDelegate(CNContactProperty property);
+        public event ContactPropertySelectedDelegate ContactPropertySelected;
+
+        internal void RaiseContactPropertySelected(CNContactProperty property)
+        {
+            if (this.ContactPropertySelected != null) this.ContactPropertySelected(property);
+        }
+        #endregion
+    }
 
 }
